+++ src/components/asset-list/RecyclerAssetList2/core/ViewTypes.ts
@@ -10,15 +10,16 @@ export enum CellType {
   SAVINGS_HEADER_SPACE_BEFORE = 'SAVINGS_HEADER_SPACE_BEFORE',
   SAVINGS = 'SAVINGS',
   POOLS_HEADER = 'POOLS_HEADER',
-  PROFILE_HEADER_AVATAR = 'PROFILE_HEADER_AVATAR',
-  PROFILE_HEADER_AVATAR_SPACE_AFTER = 'PROFILE_HEADER_AVATAR_SPACE_AFTER',
-  PROFILE_HEADER_COMPACT = 'PROFILE_HEADER_COMPACT',
-  PROFILE_HEADER_NAME = 'PROFILE_HEADER_NAME',
-  PROFILE_HEADER_NAME_SPACE_AFTER = 'PROFILE_HEADER_NAME_SPACE_AFTER',
-  PROFILE_HEADER_BALANCE = 'PROFILE_HEADER_BALANCE',
-  PROFILE_HEADER_BALANCE_SPACE_AFTER = 'PROFILE_HEADER_BALANCE_SPACE_AFTER',
-  PROFILE_HEADER_ACTION_BUTTONS = 'PROFILE_HEADER_ACTION_BUTTONS',
-  PROFILE_HEADER_ACTION_BUTTONS_SPACE_AFTER = 'PROFILE_HEADER_ACTION_BUTTONS_SPACE_AFTER',
+  PROFILE_ACTION_BUTTONS_ROW = 'PROFILE_ACTION_BUTTONS_ROW',
+  PROFILE_ACTION_BUTTONS_ROW_SPACE_AFTER = 'PROFILE_ACTION_BUTTONS_ROW_SPACE_AFTER',
+  PROFILE_AVATAR_ROW = 'PROFILE_AVATAR_ROW',
+  PROFILE_AVATAR_ROW_SPACE_BEFORE = 'PROFILE_AVATAR_ROW_SPACE_BEFORE',
+  PROFILE_AVATAR_ROW_SPACE_AFTER = 'PROFILE_AVATAR_ROW_SPACE_AFTER',
+  PROFILE_BALANCE_ROW = 'PROFILE_BALANCE_ROW',
+  PROFILE_BALANCE_ROW_SPACE_AFTER = 'PROFILE_BALANCE_ROW_SPACE_AFTER',
+  PROFILE_NAME_ROW = 'PROFILE_NAME_ROW',
+  PROFILE_NAME_ROW_SPACE_AFTER = 'PROFILE_NAME_ROW_SPACE_AFTER',
+  PROFILE_STICKY_HEADER = 'PROFILE_STICKY_HEADER',
   UNISWAP_POOL = 'UNISWAP_POOL',
   NFTS_HEADER = 'NFTS_HEADER',
   NFTS_HEADER_SPACE_BEFORE = 'NFTS_HEADER_SPACE_BEFORE',
@@ -50,7 +51,7 @@ export type CoinDividerExtraData = {
   defaultToEditButton: boolean;
 };
 export type AssetsHeaderExtraData = {
-  type: CellType.PROFILE_HEADER_COMPACT;
+  type: CellType.PROFILE_STICKY_HEADER;
   value: string;
 };
 export type PoolsHeaderExtraData = {
